net_udp.c: Fixed FIONREAD ioctl to take an int* argument instead of an unsigned long*.
git-svn-id: file:///svn/uhexen2/code/trunk@4312 eaf649e6-6c13-0410-b194-84221022858e

+++ engine/hexen2/net_udp.c
@@ -364,7 +364,7 @@ sys_socket_t UDP_CheckNewConnections (void)
 		return net_acceptsocket;
 	}
 #else
-	unsigned long	available;
+	int		available;
 	struct sockaddr_in	from;
 	socklen_t	fromlen;
 	char		buff[1];
